adding uint8

+++ Homework3/myimageprocessing.m
@@ -5,8 +5,8 @@ function count = myimageprocessing(img,t)
     I = img <= t;
     J = img > t;
     avg = [mean(img(I)), mean(img(J))];
-    img(I) = avg(1);
-    img(J) = avg(2);
+    img(I) = uint8(avg(1));
+    img(J) = uint8(avg(2));
     count = [sum(sum(I)), sum(sum(J))];
     imwrite(img, ['result_' num2str(t) '.png']);
 end
